Replace LogViewPlugin.LogReader with opencue.cuelogging.CueLogger
